FA-38 Промежуточный

+++ features/messenger/messenger-impl/src/main/java/io/fasthome/fenestram_messenger/messenger_impl/data/service/MessengerService.kt
@@ -17,7 +17,7 @@ import io.fasthome.network.model.BaseResponse
 class MessengerService(clientFactory: NetworkClientFactory) {
     private val client = clientFactory.create()
 
-    suspend fun sendMessage(id: Int, text: String, type: String): SendMessageResult {
+    suspend fun sendMessage(id: Long, text: String, type: String): SendMessageResult {
         val response: SendMessageResponse = client.runPost(
             path = "api/v1/chats/message/$id",
             body = SendMessageRequest(text, type)
@@ -44,7 +44,7 @@ class MessengerService(clientFactory: NetworkClientFactory) {
         return PostChatsMapper.responseToPostChatsResult(response)
     }
 
-    suspend fun getChatById(id: Int): GetChatByIdResult {
+    suspend fun getChatById(id: Long): GetChatByIdResult {
         val response: BaseResponse<GetChatByIdResponse> = client.runGet(
             path = "api/v1/chats/$id"
         )
